This is an aggregated commit from go-proton-api-old repo from henrybear327
Updates contained in this commit are:
Update README
Add user agent customization

Implement several drive endpoints
- MoveLink and MoveLinkReq helper functions
- CheckAvailableHashes
- GetRevisionAllBlocks
- EmptyTrash
- CreateFolderReq, RevisionMetadata, Revision helper functions

Update several drive endpoints
- CreateFile

+++ user_types.go
@@ -7,11 +7,11 @@ type User struct {
 	Email       string
 	Keys        Keys
 
-	UsedSpace uint64
-	MaxSpace  uint64
-	MaxUpload uint64
+	UsedSpace int64
+	MaxSpace  int64
+	MaxUpload int64
 
-	Credit   int
+	Credit   int64
 	Currency string
 
 	ProductUsedSpace ProductUsedSpace
